mobile: rename bundle id app.status.mobile
