+++ mobile/android/qt6/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="im.status.app"
+    package="app.status.mobile"
     android:installLocation="auto"
     android:versionCode="1"
     android:versionName="1.0">
@@ -43,7 +43,7 @@
         android:hasFragileUserData="true"
         android:fullBackupOnly="false">
         <activity
-            android:name="im.status.app.StatusQtActivity"
+            android:name="app.status.mobile.StatusQtActivity"
             android:configChanges="orientation|uiMode|screenLayout|screenSize|smallestScreenSize|layoutDirection|locale|fontScale|keyboard|keyboardHidden|navigation|mcc|mnc|density"
             android:launchMode="singleTop"
             android:resizeableActivity="false"
