+++ mobile/android/qt6/src/app/status/mobile/SecureAndroidAuthentication.java
@@ -1,4 +1,4 @@
-package im.status.app;
+package app.status.mobile;
 
 import android.app.Activity;
 import android.content.Context;
